refactor: InputContainer 분리
- taskTitle 상태와 관련된 로직 이동
- InputContiner를 위한 test 작성
- Page, Page test 삭제
