+++ src/App.test.jsx
@@ -1,12 +1,16 @@
-import { render } from '@testing-library/react';
+import { render, fireEvent } from '@testing-library/react';
 
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import App from './App';
 
 jest.mock('react-redux');
 
 test('App', () => {
+  const dispatch = jest.fn();
+
+  useDispatch.mockImplementation(() => dispatch);
+
   useSelector.mockImplementation((selector) => selector({
     tasks: [
       { id: 1, title: '아무것도 하지 않기 #1' },
@@ -19,7 +23,9 @@ test('App', () => {
   ));
 
   expect(getByText(/추가/)).not.toBeNull();
+  expect(getByText(/아무것도 하지 않기 #1/)).not.toBeNull();
+  expect(getByText(/아무것도 하지 않기 #2/)).not.toBeNull();
 
-  // TODO: 통합 테스트 코드 작성
-  // CodeceptJS => 실제 브라우저에서 사용자 테스트 실행 가능.
+  fireEvent.click(getByText(/추가/));
+  expect(dispatch).toBeCalled();
 });
