+++ src/Container/Input/InputContainer.jsx
@@ -1,0 +1,32 @@
+import { useDispatch, useSelector } from 'react-redux';
+
+import Input from './Input';
+
+import {
+  updateTaskTitle,
+  addTask,
+} from '../../store/actions';
+
+export default function InputContainer() {
+  const dispatch = useDispatch();
+
+  const { taskTitle } = useSelector((state) => ({
+    taskTitle: state.taskTitle,
+  }));
+
+  function handleChangeTitle(event) {
+    dispatch(updateTaskTitle(event.target.value));
+  }
+
+  function handleClickAddTask() {
+    dispatch(addTask());
+  }
+
+  return (
+    <Input
+      value={taskTitle}
+      onChange={handleChangeTitle}
+      onClick={handleClickAddTask}
+    />
+  );
+}
