--- src/Container/Page/Page.jsx
@@ -1,18 +1,0 @@
-import Input from '../Input/Input';
-import ListContainer from '../List/ListContainer';
-
-export default function Page({
-  taskTitle, onChangeTitle, onClickAddTask,
-}) {
-  return (
-    <div>
-      <h1>To-do</h1>
-      <Input
-        value={taskTitle}
-        onChange={onChangeTitle}
-        onClick={onClickAddTask}
-      />
-      <ListContainer />
-    </div>
-  );
-}
